backend/requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai>=1.30.0
google-generativeai==0.3.2
python-dotenv==1.0.0
anthropic>=0.39.0
groq>=0.12.0
supabase==2.7.4
